production/agent/requirements.txt
# Agent Module Dependencies

# OpenAI Agents SDK
openai>=1.10.0
openai-agents>=0.1.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
